fix: rename description.ts to validationState.ts to avoid case collision
description.ts (the validation-state helper) collided with Description.tsx
(the React description components) on case-insensitive file systems. Rename
the helper file to validationState.ts, which also better describes its
contents, and update its importers.
